face-recognition-project/requirements.txt
face-recognition==1.2.3
numpy==1.23.3
tk==0.1.0
cvzone==1.5.6
opencv-python==4.7.0.68
pillow==9.2.0
pyqt5==5.15.6